[passport] keep context cookie fix in entrypoint
